Created a new component to house logic for similar logic between update/createSubscription components. Turned update/createSubscription Components to functional components. Adding parsing for days.
